+++ src/qubic.ts
@@ -1,5 +1,6 @@
 import { QubicProvider } from "./provider";
 import { Chain } from "./chain";
+import { Transaction } from "./transaction";
 import { Identity } from "./identity";
 import { Contract } from "./contract";
 import { Utils } from './utils';
@@ -9,6 +10,7 @@ export default class Qubic {
   private provider: QubicProvider;
   public chain: Chain;
   public identity: Identity;
+  public transaction: Transaction;
   public contract: (address: string) => Contract;
   public utils = Utils;
 
@@ -16,6 +18,7 @@ export default class Qubic {
     this.provider = new QubicProvider(options);
     this.chain = new Chain(this.provider);
     this.identity = new Identity(this.provider);
+    this.transaction = new Transaction(this.provider);
     this.contract = (address: string) => new Contract(this.provider, address);
   }
 }
